refactor(presenters): complete site-to-email decoupling
Added remaining shared functions to window-helpers.ts:
- moonAstroContext
- forecastBestLine
- sessionConfidenceLabel
- sessionVolatilityLabel
- sessionRecommendationHeadline
- sessionRecommendationBody
- sessionRunnerUpLine

Updated all site sections to import from shared instead of email:
- window.ts, hero.ts, long-range.ts, spur-of-moment.ts
- forecast.ts, session-rec.ts, alternatives.ts
- hourly-outlook.ts now imports from email/outdoor-outlook-model directly

Removed duplicate function implementations from email/time-aware.ts

Relates to #127
