+++ src/presenters/shared/window-helpers.ts
@@ -6,10 +6,11 @@
  * logic for displaying photography windows, sessions, and related metadata.
  */
 
-import type { SessionId } from '../../types/session-score.js';
+import type { SessionId, SessionRecommendation, SessionRecommendationSummary } from '../../types/session-score.js';
 import type { DaySummary, Window, WindowDisplayPlan } from '../../types/brief.js';
 import { explainAstroScoreGap } from '../../domain/editorial/astro-score-explanation.js';
 import { windowRange } from '../../domain/windowing/index.js';
+import { moonIconForPct } from './brief-primitives.js';
 
 /**
  * Maps a session ID to its human-readable display name.
@@ -180,4 +181,82 @@ export function timeAwareLocalSummary(
     return `${earlier.label}: ${windowRange(earlier)} at ${earlier.peak}/100 was the strongest local window earlier today. No local photo window remains today.`;
   }
   return dedupedLines.join('\n');
+}
+
+/**
+ * Returns a moon astro context description with icon for the given moon percentage.
+ */
+export function moonAstroContext(moonPct: number): string {
+  const icon = moonIconForPct(moonPct);
+  if (moonPct <= 15) return `${icon} Dark skies — excellent for astrophotography`;
+  if (moonPct <= 40) return `${icon} Low moon glow — good for astrophotography`;
+  if (moonPct <= 70) return `${icon} Moderate moon — astrophotography compromised`;
+  if (moonPct <= 90) return `${icon} Bright moon — poor for astrophotography`;
+  return `${icon} Full moon — avoid astrophotography`;
+}
+
+/**
+ * Generates the best line for a day's forecast.
+ */
+export function forecastBestLine(day: DaySummary): string {
+  const isAstroLed = (day.astroScore ?? 0) > (day.photoScore ?? 0);
+  if (isAstroLed) {
+    return `Best local astro around ${day.bestAstroHour || 'nightfall'}`;
+  }
+  return `Best at ${day.bestPhotoHour || '-'} - ${displayBestTags(day.bestTags)}`;
+}
+
+/**
+ * Returns a label for session confidence.
+ */
+export function sessionConfidenceLabel(confidence: import('../../types/session-score.js').SessionConfidence): string {
+  switch (confidence) {
+    case 'high':
+      return 'High confidence';
+    case 'medium':
+      return 'Fair confidence';
+    case 'low':
+      return 'Low confidence';
+  }
+}
+
+/**
+ * Returns a volatility label for a session recommendation.
+ */
+export function sessionVolatilityLabel(primary: SessionRecommendation): string | null {
+  if (primary.volatility === null || primary.volatility === undefined) return null;
+  if (primary.session === 'storm') {
+    return primary.volatility >= 20
+      ? `Volatile opportunity · spread ${primary.volatility} pts`
+      : `Changeable setup · spread ${primary.volatility} pts`;
+  }
+  return `Spread ${primary.volatility} pts`;
+}
+
+/**
+ * Returns the headline for a session recommendation.
+ */
+export function sessionRecommendationHeadline(primary: SessionRecommendation): string {
+  return `${displaySessionName(primary.session)} at ${primary.hourLabel}`;
+}
+
+/**
+ * Returns the body text for a session recommendation.
+ */
+export function sessionRecommendationBody(primary: SessionRecommendation): string {
+  const reasons = primary.reasons.slice(0, 2).join(' ');
+  const warning = primary.warnings[0] || '';
+  return [reasons, warning].filter(Boolean).join(' ')
+    || `${displaySessionName(primary.session)} is the strongest specialist setup in the current forecast.`;
+}
+
+/**
+ * Returns a runner-up line for a session recommendation summary.
+ */
+export function sessionRunnerUpLine(summary: SessionRecommendationSummary | undefined): string | null {
+  const primary = summary?.primary;
+  const runnerUp = summary?.runnerUps?.[0];
+  if (!primary || !runnerUp) return null;
+  if (primary.confidence !== 'low' && (primary.volatility ?? 0) < 20) return null;
+  return `Runner-up: ${displaySessionName(runnerUp.session)} at ${runnerUp.hourLabel} (${runnerUp.score}/100).`;
 }
