+++ src/presenters/site/sections/hero.ts
@@ -5,7 +5,7 @@ import {
   type SummaryStat,
 } from '../../shared/brief-primitives.js';
 import { windowRange } from '../../../domain/windowing/index.js';
-import { moonAstroContext } from '../../email/time-aware.js';
+import { moonAstroContext } from '../../shared/window-helpers.js';
 import type { Window } from '../../../types/brief.js';
 
 function sStatGrid(items: SummaryStat[], variant: 'dark' | 'light' = 'dark'): string {
